Move Indexer to common
