+++ project/Dependencies.scala
@@ -63,7 +63,7 @@ object ExternalDependencies {
 object ServiceDependencies {
   import ExternalDependencies._
   val common: Seq[ModuleID] =
-    scalatest ++ logging ++ uPickle ++ akkaDeps
+    scalatest ++ logging ++ uPickle ++ akkaDeps ++ elasticsearchLowLevel
 
   val ingestor: Seq[ModuleID] =
     scalatest ++
@@ -72,6 +72,6 @@ object ServiceDependencies {
       Seq(akka.actorTyped, akka.stream, akka.http, akka.streamTestkit)
 
   val aggregator: Seq[ModuleID] = {
-    scalatest ++ logging ++ config ++ akkaDeps ++ elasticsearchLowLevel ++ awsLambda ++ awsSecrets
+    scalatest ++ logging ++ config ++ akkaDeps ++ awsLambda ++ awsSecrets
   }
 }
